chore: use latest svc-storage-client-grpc with prelude
