+++ client-grpc/src/prelude.rs
@@ -4,6 +4,5 @@ pub use super::client as scheduler;
 pub use super::service::Client as SchedulerServiceClient;
 
 pub use lib_common::grpc::{Client, ClientConnect, GrpcClient};
-pub use svc_storage_client_grpc as scheduler_storage;
-pub use svc_storage_client_grpc::Timestamp;
-pub use tonic::transport::Channel;
+pub use scheduler_storage::Timestamp;
+pub use svc_storage_client_grpc::prelude as scheduler_storage;
