+++ client-grpc/src/service.rs
@@ -1,4 +1,5 @@
 //! Client Library: Client Functions, Structs, Traits
+
 /// gRPC object traits to provide wrappers for grpc functions
 #[tonic::async_trait]
 pub trait Client<T>
@@ -16,19 +17,17 @@ where
     ///
     /// # Errors
     ///
-    /// Returns [`tonic::Status`] with [`Code::Unknown`](tonic::Code::Unknown) if
-    /// the server is not ready.
+    /// Returns [`tonic::Status`] with [`tonic::Code::Unknown`] if the server is not ready.
     ///
     /// # Examples
     /// ```
     /// use lib_common::grpc::get_endpoint_from_env;
-    /// use svc_scheduler_client_grpc::client::{ReadyRequest, RpcServiceClient};
+    /// use scheduler::{ReadyRequest, SchedulerClient};
     /// use svc_scheduler_client_grpc::prelude::*;
-    /// use tonic::transport::Channel;
     ///
     /// async fn example () -> Result<(), Box<dyn std::error::Error>> {
     ///     let (host, port) = get_endpoint_from_env("SERVER_HOSTNAME", "SERVER_PORT_GRPC");
-    ///     let connection = GrpcClient::<RpcServiceClient<Channel>>::new_client(&host, port, "scheduler");
+    ///     let connection = SchedulerClient::new_client(&host, port, "scheduler");
     ///     let response = connection
     ///         .is_ready(ReadyRequest {})
     ///         .await?;
