+++ server/Cargo.toml
@@ -69,7 +69,7 @@ tag = "v0.1.1-develop.5"
 [dependencies.svc-storage-client-grpc]
 features = ["flight_plan", "itinerary", "vehicle", "vertipad", "vertiport"]
 git      = "https://github.com/Arrow-air/svc-storage.git"
-tag      = "v0.10.1-develop.20"
+tag      = "v0.10.1-develop.21"
 
 [dependencies.log4rs]
 features = [
